clean up code and create a spreate component for quote
