+++ components/Quote/Quote.style.js
@@ -1,0 +1,15 @@
+import styled from "styled-components/native";
+
+const Container = styled.View`
+  flex: 1;
+  background-color: papayawhip;
+  justify-content: center;
+  align-items: center;
+`;
+
+const Title = styled.Text`
+  font-size: 24px;
+  color: palevioletred;
+`;
+
+export { Container, Title };
